Return `NONE` argument instead of `none()` value
`parse_arg` now returns an `arg_t` with a type of `NONE` when there is no
argument present. Until now, it returned a `none()`-wrapped value, with caused
a panic when unwrapping it in `parse`.

TODO: the usage of the `maybe(arg_t)` type should be completely removed from
the code.